Add transaction events to order history
